requirements.txt
pypdf
langchain
chromadb
openai
sentence-transformers
umap-learn
tqdm
numpy
matplotlib